--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/ValidationQuery.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/ValidationQuery.java
@@ -83,7 +83,7 @@ public class ValidationQuery {
 				propertyShapeWithValue = false;
 				valueIndex = variables.size();
 				assert constraintComponent == null
-						|| !constraintComponent.getConstraintComponent().producesValidationResultValue();
+						|| !constraintComponent.getConstraintComponent().alwaysProducesValidationResultValue();
 			}
 		} else {
 			targetIndex = variables.size() - 1;