--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/planNodes/BulkedExternalLeftOuterJoin.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/planNodes/BulkedExternalLeftOuterJoin.java
@@ -21,6 +21,7 @@ import org.eclipse.rdf4j.model.Resource;
 import org.eclipse.rdf4j.query.BindingSet;
 import org.eclipse.rdf4j.query.Dataset;
 import org.eclipse.rdf4j.query.algebra.TupleExpr;
+import org.eclipse.rdf4j.query.algebra.evaluation.iterator.PeekMarkIterator;
 import org.eclipse.rdf4j.sail.SailConnection;
 import org.eclipse.rdf4j.sail.memory.MemoryStoreConnection;
 import org.eclipse.rdf4j.sail.shacl.ast.SparqlFragment;
@@ -64,13 +65,13 @@ public class BulkedExternalLeftOuterJoin extends AbstractBulkJoinPlanNode {
 			ArrayDeque<ValidationTuple> left;
 			ArrayDeque<ValidationTuple> right;
 
-			private CloseableIteration<? extends ValidationTuple> leftNodeIterator;
+			private PeekMarkIterator<? extends ValidationTuple> leftNodeIterator;
 
 			@Override
 			protected void init() {
 				left = new ArrayDeque<>(BULK_SIZE);
 				right = new ArrayDeque<>(BULK_SIZE);
-				leftNodeIterator = leftNode.iterator();
+				leftNodeIterator = new PeekMarkIterator<>(leftNode.iterator());
 			}
 
 			private void calculateNext() {
@@ -80,6 +81,15 @@ public class BulkedExternalLeftOuterJoin extends AbstractBulkJoinPlanNode {
 				}
 
 				while (left.size() < BULK_SIZE && leftNodeIterator.hasNext()) {
+					if (!left.isEmpty()) {
+						ValidationTuple peek = leftNodeIterator.peek();
+						if (peek.sameTargetAs(left.getFirst())) {
+							// stop if we detect a duplicate target since we only support distinct targets on the left
+							// side of the join
+							assert false : "Current and next left target is the same: " + peek + " " + left.getFirst();
+							break;
+						}
+					}
 					left.addFirst(leftNodeIterator.next());
 				}