--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/targets/EffectiveTarget.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/targets/EffectiveTarget.java
@@ -401,6 +401,10 @@ public class EffectiveTarget {
 		return Objects.requireNonNull(optional, "Optional was null").var;
 	}
 
+	public int size() {
+		return chain.size();
+	}
+
 	public enum Extend {
 		left,
 		right