--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/targets/TargetClass.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/targets/TargetClass.java
@@ -66,6 +66,7 @@ public class TargetClass extends Target {
 			Resource clazz = targetClass.stream().findAny().get();
 			planNode = new UnorderedSelect(connection, null, RDF.TYPE, clazz,
 					dataGraph, UnorderedSelect.Mapper.SubjectScopedMapper.getFunction(scope), null);
+			return planNode;
 		} else {
 			planNode = new Select(connection,
 					SparqlFragment.bgp(Set.of(),