… core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/constraintcomponents/SparqlConstraintComponent.java
… core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/planNodes/FilterByPredicateObject.java
… core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/planNodes/PatternFilter.java
… core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/planNodes/SparqlConstraintSelect.java
… core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/results/ValidationResult.java
… core/sail/shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/AbstractShaclTest.java
… core/sail/shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/ShaclTestWithoutSparqlValidationTest.java
… core/sail/shacl/src/test/resources/test-cases/class/complexPath/invalid/case7/query1.rq
… core/sail/shacl/src/test/resources/test-cases/class/complexPath/invalid/case7/report.ttl
… core/sail/shacl/src/test/resources/test-cases/class/complexPath/shacl.trig
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case1/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case1/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case2/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case2/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case3/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case3/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case4/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case4/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case5/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case5/query2.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case5/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case6/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case6/query2.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/invalid/case6/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/shacl.trig
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/valid/case1/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/valid/case1/report.ttl
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/valid/case2/query1.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/valid/case2/query2.rq
… core/sail/shacl/src/test/resources/test-cases/datatype/nestedNode/valid/case2/report.ttl