@hmottestad 2024-08-27
## Before
```
```


## After
```
Benchmark                                                          Mode  Cnt     Score      Error  Units
ComplexLargeBenchmark.disabledValidationSail                       avgt    5   332.216 ±   54.691  ms/op
ComplexLargeBenchmark.disabledValidationTransaction                avgt    5   329.656 ±   13.421  ms/op
ComplexLargeBenchmark.noPreloading                                 avgt    5   652.694 ±   61.175  ms/op
ComplexLargeBenchmark.noPreloadingBulk                             avgt    5   631.931 ±   28.237  ms/op
ComplexLargeBenchmark.noPreloadingBulkParallelCached               avgt    5   583.691 ±   68.283  ms/op
ComplexLargeBenchmark.noPreloadingNonEmpty                         avgt    5  3751.262 ±  182.784  ms/op
ComplexLargeBenchmark.noPreloadingNonEmptyParallel                 avgt    5  2469.678 ±  302.404  ms/op
ComplexLargeBenchmark.noPreloadingNonEmptyParallelLmdb             avgt    5  9497.425 ± 1303.038  ms/op
ComplexLargeBenchmark.noPreloadingNonEmptyParallelNativeStore      avgt    5  7256.163 ±  365.693  ms/op
ComplexLargeBenchmark.noPreloadingNonEmptyParallelReadCommitted    avgt    5  4295.821 ±  431.832  ms/op
ComplexLargeBenchmark.noPreloadingNonEmptyParallelRemoved          avgt    5  2755.962 ±  538.590  ms/op
ComplexLargeBenchmark.noPreloadingNonEmptyRemoved                  avgt    5  3915.580 ±  193.414  ms/op
ComplexLargeBenchmark.noPreloadingParallel                         avgt    5   677.575 ±  129.554  ms/op
ComplexLargeBenchmark.noPreloadingParallelNoCache                  avgt    5   644.252 ±   45.313  ms/op
ComplexLargeBenchmark.noPreloadingRevalidate                       avgt    5   568.130 ±   29.726  ms/op
ComplexLargeBenchmark.noPreloadingRevalidateLowMem                 avgt    5   642.213 ±   35.515  ms/op
ComplexLargeBenchmark.noPreloadingRevalidateNativeStore            avgt    5  3466.532 ±  259.464  ms/op
ComplexLargeBenchmark.noPreloadingTransactionalValidationLimit     avgt    5   970.873 ±   66.323  ms/op
ComplexLargeBenchmark.shaclNothingToValidateTransactionsPreloaded  avgt    5     0.109 ±    0.004  ms/op
```