@hmottestad 2024-08-28
Performance degradation seems to be related to TargetClass and the removal of a call to Unique there. 

It might not be the Unique, but rather that we were caching the unique and sorted data, while we now just cache the raw unsorted data.