@hmottestad 2024-08-28
Maybe we should try to optimize the bulk leftOuter/inner joins to not require the data be sorted.